* Merged HEAD to cf (revisions 2272 to 2394).

+++ src/plugins/codecompletion/classbrowserbuilderthread.cpp
@@ -1,16 +1,20 @@
+#include <sdk.h>
 #include "classbrowserbuilderthread.h"
+#include <globals.h>
 
 ClassBrowserBuilderThread::ClassBrowserBuilderThread(Parser* parser,
                                                     wxTreeCtrl& tree,
                                                     const wxString& active_filename,
                                                     BrowserOptions options,
-                                                    TokensTree* pTokens)
+                                                    TokensTree* pTokens,
+                                                    ClassBrowserBuilderThread** threadVar)
     : wxThread(wxTHREAD_DETACHED),
     m_pParser(parser),
     m_Tree(tree),
     m_ActiveFilename(active_filename),
     m_Options(options),
-    m_pTokens(pTokens)
+    m_pTokens(pTokens),
+    m_ppThreadVar(threadVar)
 {
     //ctor
 }
@@ -23,6 +27,8 @@ ClassBrowserBuilderThread::~ClassBrowserBuilderThread()
 void* ClassBrowserBuilderThread::Entry()
 {
     BuildTree();
+    if (m_ppThreadVar)
+        *m_ppThreadVar = 0;
     return 0;
 }
 
@@ -30,11 +36,25 @@ void ClassBrowserBuilderThread::BuildTree()
 {
     wxCriticalSectionLocker lock(s_MutexProtection);
 
+    if (TestDestroy())
+        return;
+
+    wxArrayString treeState;
+    wxTreeItemId root = m_Tree.GetRootItem();
+    if (root.IsOk())
+        ::SaveTreeState(&m_Tree, root, treeState);
+
 	m_Tree.Freeze();
     m_Tree.DeleteAllItems();
     TokenFilesSet currset;
     currset.clear();
     Token* token = 0;
+
+    if (TestDestroy())
+    {
+        m_Tree.Thaw();
+        return;
+    }
 
     // "mark" tokens based on scope
     bool fnameEmpty = m_ActiveFilename.IsEmpty();
@@ -48,7 +68,13 @@ void ClassBrowserBuilderThread::BuildTree()
         }
     }
 
-	wxTreeItemId root = m_Tree.AddRoot(_("Symbols"), PARSER_IMG_SYMBOLS_FOLDER);
+    if (TestDestroy())
+    {
+        m_Tree.Thaw();
+        return;
+    }
+
+	root = m_Tree.AddRoot(_("Symbols"), PARSER_IMG_SYMBOLS_FOLDER);
 	if (m_Options.viewFlat)
 	{
         TokenIdxSet::iterator it,it_end;
@@ -61,6 +87,12 @@ void ClassBrowserBuilderThread::BuildTree()
             if(!token || !token->m_IsLocal || token->m_ParentIndex!=-1 || !token->MatchesFiles(currset))
                 continue;
             AddTreeNode(root, token);
+
+            if (TestDestroy())
+            {
+                m_Tree.Thaw();
+                return;
+            }
         }
 		m_Tree.SortChildren(root);
 	}
@@ -68,10 +100,25 @@ void ClassBrowserBuilderThread::BuildTree()
 	{
         wxTreeItemId globalNS = m_Tree.AppendItem(root, _("Global namespace"), PARSER_IMG_NAMESPACE);
         AddTreeNamespace(globalNS, 0,currset);
+        if (TestDestroy())
+        {
+            m_Tree.Thaw();
+            return;
+        }
         BuildTreeNamespace(root, 0,currset);
+        if (TestDestroy())
+        {
+            m_Tree.Thaw();
+            return;
+        }
 	}
 
-	m_Tree.Expand(root);
+    if (root.IsOk())
+    {
+        ::RestoreTreeState(&m_Tree, root, treeState);
+        if (!m_Tree.IsExpanded(root))
+            m_Tree.Expand(root);
+    }
 	m_Tree.Thaw();
 	// wxString memdump = m_pTokens->m_Tree.Serialize();
 	// Manager::Get()->GetMessageManager()->DebugLog(memdump);
@@ -94,17 +141,24 @@ void ClassBrowserBuilderThread::BuildTreeNamespace(const wxTreeItemId& parentNod
         parentidx = parent->GetSelf();
     }
 
+    bool hasCurrset = currset.size() != 0;
 	for(;it != it_end; it++)
 	{
 	    Token* token = m_pTokens->at(*it);
-	    if(!token || /* !token->m_Bool || */ !token->m_IsLocal || token->m_TokenKind != tkNamespace)
+	    if(!token || /* !token->m_Bool || */ (!hasCurrset && !token->m_IsLocal) || token->m_TokenKind != tkNamespace)
             continue;
-        if(currset.size() && !token->MatchesFiles(currset))
-            continue;
+//        if(hasCurrset && !token->MatchesFiles(currset))
+//            continue;
+//        Manager::Get()->GetMessageManager()->DebugLog(_T("  + Matching namespace: ") + token->m_Name);
         ClassTreeData* ctd = new ClassTreeData(token);
         wxTreeItemId newNS = m_Tree.AppendItem(parentNode, token->m_Name, PARSER_IMG_NAMESPACE, -1, ctd);
         BuildTreeNamespace(newNS, token, currset);
         AddTreeNamespace(newNS, token, currset);
+        // remove branch if empty
+        if (!m_Tree.ItemHasChildren(newNS))
+            m_Tree.Delete(newNS);
+        if (TestDestroy())
+            return;
 	}
     m_Tree.SortChildren(parentNode);
 }
@@ -133,10 +187,14 @@ void ClassBrowserBuilderThread::AddTreeNamespace(const wxTreeItemId& parentNode,
 	wxTreeItemId node_others;
 	wxTreeItemId* curnode = 0;
 
+    bool hasCurrset = currset.size() != 0;
 	for(;it != it_end; it++)
 	{
+        if (TestDestroy())
+            return;
+
 	    Token* token = m_pTokens->at(*it);
-	    if(!token || /* !token->m_Bool || */ !token->m_IsLocal)
+	    if(!token || /* !token->m_Bool || */ (!hasCurrset && !token->m_IsLocal))
             continue;
         if(currset.size() && !token->MatchesFiles(currset))
             continue;
@@ -185,10 +243,16 @@ void ClassBrowserBuilderThread::AddTreeNamespace(const wxTreeItemId& parentNode,
 	}
     if(has_classes)
         m_Tree.SortChildren(node_classes);
+    if (TestDestroy())
+        return;
     if(has_enums)
         m_Tree.SortChildren(node_enums);
+    if (TestDestroy())
+        return;
     if(has_preprocessor)
         m_Tree.SortChildren(node_preprocessor);
+    if (TestDestroy())
+        return;
     if(has_others)
         m_Tree.SortChildren(node_others);
 }
